+++ src/main/java/com/example/secondlife/domain/like/post/dto/PostLikeCountResponse.java
@@ -1,4 +1,4 @@
-package com.example.secondlife.domain.likes.post.dto;
+package com.example.secondlife.domain.like.post.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,10 +7,9 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 @Builder
-public class PostLikeResponse {
+public class PostLikeCountResponse {
 
     private Long postId;
-    private Long userId;
-    private Long postLikeId;
+    private Long count;
 
 }
