+++ src/main/java/com/example/secondlife/domain/like/post/dto/PostLikeResponse.java
@@ -1,4 +1,4 @@
-package com.example.secondlife.domain.likes.post.dto;
+package com.example.secondlife.domain.like.post.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,9 +7,10 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 @Builder
-public class PostLikeCountResponse {
+public class PostLikeResponse {
 
     private Long postId;
-    private Long count;
+    private Long userId;
+    private Long postLikeId;
 
 }
